share code between mlb/nfl in new 'sports/games.js' file that is required from both of those spices
